Migrated handler tests to EditBillableWeight component
