+++ src/components/Office/BillableWeight/ShipmentCard/ShipmentCard.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, screen, fireEvent } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
 
 import ShipmentCard from './ShipmentCard';
 
@@ -54,42 +54,5 @@ describe('ShipmentCard', () => {
     // addresses
     expect(screen.getByText(formatAddressShort(defaultProps.pickupAddress))).toBeInTheDocument();
     expect(screen.getByText(formatAddressShort(defaultProps.destinationAddress))).toBeInTheDocument();
-  });
-
-  it('calls editEntity when the user saves', () => {
-    const tomorrow = new Date();
-    tomorrow.setDate(tomorrow.getDate() + 1);
-    const editEntity = jest.fn();
-    const defaultProps = {
-      billableWeight: 4014,
-      maxBillableWeight: 0,
-      dateReweighRequested: new Date().toISOString(),
-      departedDate: tomorrow.toISOString(),
-      pickupAddress: {
-        city: 'Rancho Santa Margarita',
-        state: 'CA',
-        postal_code: '92688',
-      },
-      destinationAddress: {
-        city: 'West Springfield Town',
-        state: 'MA',
-        postal_code: '01089',
-      },
-      estimatedWeight: 5000,
-      originalWeight: 4300,
-      reweighRemarks: 'Unable to perform reweigh because shipment was already unloaded',
-      editEntity,
-    };
-
-    render(<ShipmentCard {...defaultProps} />);
-    const newBillableWeight = 5000;
-    const newBillableWeightJustification = 'some remarks';
-    fireEvent.click(screen.getByRole('button', { name: 'Edit' }));
-    fireEvent.change(screen.getByTestId('textInput'), { target: { value: newBillableWeight } });
-    fireEvent.change(screen.getByTestId('remarks'), { target: { value: newBillableWeightJustification } });
-    fireEvent.click(screen.getByRole('button', { name: 'Save changes' }));
-    expect(editEntity.mock.calls.length).toBe(1);
-    expect(editEntity.mock.calls[0][0].billableWeight).toBe(newBillableWeight);
-    expect(editEntity.mock.calls[0][0].billableWeightJustification).toBe(newBillableWeightJustification);
   });
 });
